move admin routes to separate file and update swagger
