+++ src/main/scala/org/broadinstitute/dsde/workbench/sam/api/UserInfoDirectives.scala
@@ -57,5 +57,4 @@ trait UserInfoDirectives {
         else r
       }
     }
-
 }
